► Changes ►​ When warps are loaded, the description is trimmed of whitespaces and extra spaces are replaced with a single space. This will keep things pretty and consistent. Polished the temp fix from the hotfix resolving the RTP issue even further.
► Bug Fixes ►​
Fixed an issue involving re-loading the plugin live. The custom world-guard flag has been wrapped in a "try/catch" to allow plug-man and similar plugins to reload it.
Fixed an issue involving the MySQL support causing the plugin to thing YAML support is still a thing and failing to load warps due to it.

Signed-off-by: BuildTools <unconfigured@null.spigotmc.org>