The plugin now uses Spigot as its base. CraftBukkit/Bukkit is no longer supported. In addition, 1.16 should work out of the box as long as the classpath doesn't change outside of normal NMS.
When the “/spawn” command is used, the sender will no longer receive two messages.

When calculating a player's personal warp limit, invalid value entered as the suffix will now be skipped.

Implemented a simple check to see if the player is in bed when interacting with the GUI. There is apparently a GUI duplication glitch found in vanilla mc causing this, so this was added just in case!

Warp usage costs now go to the warp owner, if exists.

Random teleportation now attempts to retrieve a chunk when finding locations, if the chunk takes too long to obtain by the time the check comes across, the attempt will now be considered a failure. It is HIGHLY recommended that Paper Spigot is used for the best random teleportation experience. If not, please pre-generate your world as the plugin will NOT play nicely and generate it for you ;).

Random teleportation now randomizes using the smart limit (starting at zero) subtracted from the configured bounds size. The smart limit will increase by 0.5% of the bounds size for every failed attempt at finding a safe location. The system now also uses the full scope of the bounds size,  meaning there is a possibility for the player to mean the world's limit (higher the size, lower the chance).

Random teleportation will no longer teleport a player to a location that has a distance less than 10% the current smart bounds radius from base-location. This just forces the system to push players out a little more based on smart bounds. As an example, 10% of a 10,000 bounds radius will result in 1,000 blocks from the base-location minimum.

Warp sorting now compares traffic (if featured mode is enabled), the ratio of likes to dislikes, and now sorts alphabetically before analyzing furthur. This will make GUI interactions feel more consistent rather than getting a random list every time due to warps having identical information.

Fixed an issue where the “/warps <name>” would not charge players in the same manner as other features.

Fixed an issue where 1.12.2 and under paper spigot users would encounter errors when the async chunk load methods are used.

Fixed an issue where the “hyperdrive.like” permission was not nested under the global “hyperdrive.use.*” permission as a child.

Fixed an issue where the spawn location would load using integers of the actual saved values. This causes the spawn location to be in-accurate and teleports the player to the corner of the block at x and z equals zero.

Fixed an issue where the chat help menu’s hover text always stated the word “administrator”.

Fixed an issue involving MySQL. The properties section found in the “mysql-connection” section has been reformatted to be a completely customizable extension string.

Signed-off-by: BuildTools <unconfigured@null.spigotmc.org>